requirements.txt
python=3.12.5
matplotlib
numpy
pandas
tensorflow
pillow